test register handler

+++ datastore/interface.go
@@ -9,6 +9,6 @@ import (
 // CustomerRepository is an interface for customer storage
 type CustomerRepository interface {
 	Save(ctx context.Context, customer *customer.Customer) error
-	FindByID(ctx context.Context, ID uint) (*customer.Customer, error)
+	FindByID(ctx context.Context, ID uint64) (*customer.Customer, error)
 	FindByEmail(ctx context.Context, email string) (*customer.Customer, error)
 }
